Minor changes
Changed Default window location for the two forms to be center of
screen.  This makes the error dialog in Form2's position suck less since
it always appears in the center of the screen (apparently, you can't
center a dialog box to it's parent in C#???(???(!)))
Also fixed some weirdness with the combobox's values being populated on
only the values from the DB.  I think I forgot to change this?  Or was
it supposed to be like that?  Anyway, it doesn't make sense so I put it
back to how it was originally being done on Form2 (hard coded values
ftw...).